[libngf-qt] Avoid emitting playing state twice.
When calling the "play" method over DBus, if the
DBus reply is positive, the "playing" state is
emitted. Then the ngf daemon is emitting a state
change to "playing" which is transmitted by the
client. This commit is avoiding to transmit this
latter "playing" state in the normal case.

If the daemon is unable to play the effect, then
the "playing" signal is emitted (because the DBus
call succeed) and is then followed by a "failed"
signal, thanks to the state change broadcasted by
ngfd. This commit is not changing this behaviour.