[libngf-qt] Simplify d-bus client by avoiding startup round-trips. JB#62786
The initial setup was able to get the process blocking if ngfd
was being unresponsive. Furthermore the name owner fetching
on startup and QDBusInterface creation both do a round-trip.

We don't really need to know at startup whether NGFD is there to serve.
We can just fire up events and they will end up as eventFailed
if there's some problem. If needed, the API can be augmented to separate
the ngfd not running case, but so far there seems quite little interest
in client side whether the play requests succeed.